backend/requirements.txt
# FastAPI currently resolves Starlette 0.48.x; API-level Range headers are blocked in app middleware.
fastapi==0.120.0
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.5.2
# python-jose pulls rsa/pyasn1 transitively; no conflict-free pyasn1 fix pin is available with this stack yet.
python-jose[cryptography]==3.4.0
passlib==1.7.4
email-validator==2.2.0
python-multipart==0.0.22
pypdf==6.9.2
openpyxl==3.1.5
openai==1.51.2
httpx==0.27.2
